backend/pyproject.toml
[tool.poetry]
name = "metrics-exercise"
version = "0.1.0"
description = ""
authors = ["Cristóbal Contreras Rubio <crisconru@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.74.1"
uvicorn = "^0.17.5"
SQLAlchemy = "^1.4.31"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^4.0.1"
mypy = "^0.931"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
